delligent/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
firebase-admin==6.4.0
pinecone-client==3.0.0
google-generativeai==0.3.2
python-dotenv==1.0.1
pydantic==2.5.3
python-multipart==0.0.6

